working

+++ src/CrossSection.cpp
@@ -117,7 +117,7 @@ void CrossSection::updateMacroXS(const int& l, float* dnst)
 void CrossSection::updateMacroXS(float* dnst)
 {
 #pragma omp parallel for
-	for (size_t l = 0; l < _nxyz; l++)
+	for (int l = 0; l < _nxyz; l++)
 	{
 		updateMacroXS(l, dnst);
 	}
@@ -201,7 +201,7 @@ void CrossSection::updateXS(const int& l, const float* dnst, const float& dppm, 
 void CrossSection::updateXS(const float* dnst, const float* dppm, const float* dtf, const float* dtm)
 {
 #pragma omp parallel for
-	for (size_t l = 0; l < _nxyz; l++)
+	for (int l = 0; l < _nxyz; l++)
 	{
 		updateXS(l, dnst, dppm[l], dtf[l], dtm[l]);
 	}
